Increase IBUF_SIZE to prevent deadlock
IBUF_SIZE needs to be at least as large as the number of sourceIds
allocated for instruction fetch.

Since ibuffer is instantiated per warp however, and there is no
contraint on how many of the outstanding ifetch requests map to a
specific warp, we need to accommodate the worst case where all requests
originate from a single warp by setting the per-warp ibuffer size to
nSrcIds.  The current number is pretty wasteful when multiplied by
NUM_WARPS.